add workouts context nulling

+++ frontend/src/components/WorkoutDetails.jsx
@@ -3,16 +3,24 @@ import {useWorkoutsContext} from "../hooks/useWorkoutsContext"
 //date-fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNowStrict'
 
+import { useAuthContext } from "../hooks/useAuthContext"
 
 function WorkoutDetails({workout}) {
 
   //to grab the dispatch function from the useWorkoutContext
   const {dispatch}=useWorkoutsContext()
 
-
+  const {user}=useAuthContext()
 const handleClick=async()=>{
+  if(!user){
+    return;
+  }
+
     const response=await fetch('http://localhost:5001/api/workouts/'+workout._id, {
-      method:'DELETE'
+      method:'DELETE',
+      headers:{
+        'Authorization':`Bearer ${user.token}`
+      }
     })
 
     const json=await response.json()
